pyproject.toml
[project]
name = "arbitrage-trading"
version = "0.1.0"
description = "암호화폐 거래소 간 스프레드 모니터링 대시보드"
authors = [
    { name = "finai6" }
]
dependencies = [
    "streamlit>=1.32.0",
    "pandas>=2.2.0",
    "plotly>=5.18.0",
    "ccxt>=4.2.15",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=20.7",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ruff>=0.1.6",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3 